refactor supervised tuning into generative_model module, PR feedback, tests
