+++ docs/apache-airflow-providers-google/operators/cloud/vertex_ai.rst
@@ -615,11 +615,7 @@ The operator returns the model's response in :ref:`XCom <concepts:xcom>` under `
     :start-after: [START how_to_cloud_vertex_ai_generative_model_generate_content_operator]
     :end-before: [END how_to_cloud_vertex_ai_generative_model_generate_content_operator]
 
-
-Performing Supervised Fine Tuning
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-
-To train and deploy a tuned generative model to an endpoint you can use
+To run a supervised fine tuning job you can use
 :class:`~airflow.providers.google.cloud.operators.vertex_ai.supervised_fine_tuning.SupervisedFineTuningTrainOperator`.
 The operator returns the tuned model's endpoint name in :ref:`XCom <concepts:xcom>` under ``tuned_model_endpoint_name`` key.
